Initial commit
A long, long time ago, software was written in one go. Modifications
after the initial release were scare, and to this date, `ed` remains
mostly unchanged since the year of 1971.

However, the growing complexity of software lead to continuous
modifications, which in turn called for tools to manage those
changes to the code, and version control systems were born. Soon,
software was not crafted by a single engineer any more, but by a team
of developers. Suddenly, the early systems operating on a single
machine's disk was insufficient for the task, and systems operating as
client and server were introduced.

Somewhere along the way, branching and merging became a necessity,
helping developers not to stand on each others' toes.

The locking fashion and the single point of failure the server-centered
approach posed proved to be too much of a drawback especially for those
who crafted huge systems, with thousands of change-sets a day. The time
was right for distributed version control systems.

One developer started out just like some of his colleagues 30 years
earlier, with bare files. At first, he committed the changes he made by
hand in a simple data-storage based on raw files and hashes. Soon, the
system could be used to version control itself and we ended up with the
system we are using right now: git.

Distributed version control systems are now the mainstream. However,
issue trackers remain centralized and server-oriented. Project
maintainers are either stuck with platforms set up years, sometimes
decades, ago or have to manage and unify the many issue- and
bug-trackers used. Although some distributed issue tracker were written,
none seems to be fit for the task and many have been abandoned already.

Unsatisfied with the systems presented to us and unimpressed with the
failures of others, we set out to an epic journey. Like Linus Torvalds
did not rely on another version control system when writing git, we will
refuse to use an existing issue tracker for the development of our
"git-dit". Maybe this will end up as another dead project at some point,
but we are confident that it is time for distributed issue trackers to
be finally used.

Watch out for empty commits, because this repo will have many of them.