Create a helper for posting new issues/comments
We need tooling for simplifying the process of posting issues and/or
comments. The tool should take as an input argument the parent, which
may or may not be a post. The parent may be optional, in which case a
missing parent could indicate that the post is a new issue, without any
parent commit. Optionally, such behavior could be made available
through some switch.

It may make sense to actually implement two separate tools, one for
posting issues, and one to reply to issues and comments. This would not
require the tool to determine whether the parent already belongs to an
issue or whether it is a "regular" commit.

A tool should behave like existing git commands, in terms of how the
commit message is read from the user. I'm not yet sure whether this
should be a plumbing command or whether this might also be used by
users. In the latter case, we should spawn an editor by default, in the
former we read from std-in. In any case, the established `-F` and `-m`
arguments should be supported at some point.

A possible implementation could use `git commit-tree`. This command
expects as argument a tree object to be added to the commit. An empty
tree can apparently be generated using

    git hash-object -t tree /dev/null

`git commit-tree` prints the hash of the commit created, which should be
used to create the refs, e.g. using `git update-ref`.