Re: Decide on how to name refs
So, we kept the current `refs/dit/<hash>/init` ref as a delimiter for
when we iterate over the comments, e.g. to aggregate meta-data.

The thing is, the head and leaves already have the hash in the ref's
name. We can just extract and use that hash when we need it, and remove
the `/init` ref as soon as we have either `/head` or leaves.

On second thought, we could omit the `/init` ref altogether and just
start out with a `/head`. I'm implementing this right away...