Introduce git-dit-create-message script
This commit introduces the git-dit-create-message command script, which
is a plumbing command.

It can be used to create a new message (u no say).

A new message can be either a completely new message or a reply to an
existing message or a message which is attached to an existing commit,
for example for reporting a bug which is introduced by a specific
commit.

Here it all begins...