Re: Create a helper for posting new issues/comments
I have created a first draft (with a lot TODOs) in

    b1262840fd5f87bafa1d7e64807f7eddc382620d

which is the HEAD of the branch

    m/add-comment-adder

I would like to implement this as plumbing
command which does _only_ add comments on existing ones.

This is _not_ runnable by now, which is also why it is not checked in as
executable script.

Feel free to review and comment.