Using an empty tree for all new issue commits might be wrong
Creating a "regular" empty commit does not result in a commit having an
empty tree but in a commit which has the same tree as the parent.
Creating a commit with an empty tree on top of a commit with a non-empty
tree has the same semantics as removing all the contents from the (root)
directory.

Hence, if we create a bug-report on top of a regular commit, we may run
into problems when we try to include patches. This is incompatible with
some of the possible modes of operation conceivable. While we could just
discourage or forbid users from using such workflows, using the parent's
tree for non-initial "issue commits" would be more natural.