Make "git dit" actually work
Because of an error on my part, "git dit" was introduced in an
incomplete state, making it utterly useless.

For git dit to work, it must be executable. Also, git doesn't look for
scripts with the "sh" extension when invoking subcommands, so naming the
thing "git-dit.sh" just doesn't make sense.