Implement meta-data
We need a mechanism to add (structured) meta-information to an issue, that is:

    * Status
    * Tags
    * Assignee
    * Type (E.G.: feature-request, bug-report, ...)
    * Prio
    * Severity

of an issue.
The meta-information should be key-value-pairs.

The key should be prefixed with something like "Dit-Issue-" or similar, to be
sure we do not run into issues with other tools, etc.
(Lets start the bike-shedding on CamelCase vs. lowercase)

The above list is a proposed set of default meta information. The mechanism
should be implemented using git-commit-tags. It should be extensible for other
types of meta-information.

We need:

    * A way to add meta-information at issue-creation
    * A way to add/alter meta-information of an existing issue
    * A way to list meta-information of an issue

Also, it would be neat to extend `list-issues` to be able to filter for
issue-metadata, for example "list all issues containing tag=foo".

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>