Fix: Eliminate duplicated lines, sort output, more fancy format.
This patch fixes the issue that output may be contain duplicated
entries.

The approach is simple: Prefix the output with the commit-date
(timestamp), which is later filtered out.
Sort the lines, reverse them, make them unique.

Let the output also contain the date (in nice human readable form: "2
weeks ago") so we have a nice listing of issues.