Use `COMMIT_EDITMSG` by default, instead of a temporary file
We currently use a temporary file by default. However, users kind-of
expect commit messages to be edited in `.git/COMMIT_EDITMSG`. Users of
vim will expect a certain "syntax-highlighting" for commit messages,
which relies on the file being edited to be named `COMMIT_EDITMSG`.

Matthias pointed out that some users might have strange workflows,
potentially leading to conflicts. If a user opens an issue while writing
a commit, for example, things break. To support such workflows, we could
introduce a flag or configuration parameter for using a temporary or
predefined file.