Invoke the pager only once
`git dit list -l` invokes `git dit show` which, in turn, invokes
`git log`. Previously, if a pager was configured, that pager was invoked
with every single call to `git log`.

This command solves this issue by invoking `git dit show` with the
`--no-pager` option and piping the output to the appropriate pager
(which is `cat` if the user did specify `--no-pager` when listing the
issues).