GC functionality
We want a GC-like functionality that automatically cleans up
dead/outdated references and so on.

I would implement this as a set of tools, actually, which can be called
one by one by a "top level" git-dit-gc tool, as the individual tools may
be rather complex in what they do compared to the tools we currently
have.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
Dit-status: open
Dit-priority: high