Functionality to execute dit commands in distant repo
A nice thing about dit is that we can build workflows where more than
one repository is involved in the project.

For example, one can use one repository for source code, and another
repository for issues.
If people would like to have both in one repository, even the
possibility of having them in the same repository, but in two different
remote repositories is given (in fact I would even make this the
suggested default setup).

This issue is about another thing, though: If one wants to have two
repositories, one for code, the other for issues, we should have a
possibility to execute all `git-dit` commands in the other repository.

Consider the following scenario:

    /home/user/project_a
    /home/user/project_a_issues

If the user works with his project, he may want to execute `git dit
list` or something like that. We should have a possibility to
automatically execute this in /home/user/project_a_issues, even if PWD
is /home/user/project_a.

A configuration setting for this would be really neat. Tools like `tig`,
`git log` and so on naturally wouldn't list the issues from the other
repository (of course), so even more convenience here.

So I propose to have a configuration option to do work in a distant
repository for all `git-dit-*` commands. Of course this is low-prio.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
Dit-status: open
Dit-priority: low
Dit-type: idea