Issue Namespaces
A idea would be to introduce namespaces for issues.

That means, that a issue ref could have an (optional?) namespace part:

    refs/dit/<namespace>/<hash>...

where <namespace> is a normal name with optional sub-namespaces, E.G.:

    refs/dit/backend/database/<hash>...

This would give us one really neat thing: with tools like gitolite, one
can define sub-maintainers and allow them to push to certain namespaces
only.

Things that we should implement then:

1. Moving things between namespaces - complicated from a data storage
point of view, as moving alters hashes

2. Options in list/new/... other tools to work with a certain namespace.

Discussion is also required in how to handle integrations, then. This
may even yield this feature useless, as we would need to automatically
map (for example) emails to a certain namespace.

Anyways, this issue is a mental note that we might consider some
discussion here and it might not be implemented until a hand full of
versions of this tool are published.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
Dit-status: open
Dit-priority: low
Dit-type: idea