Allow replies to have additional references
A feature which we often talked about was the ability to use additional
parents to reference commits from issues-messages. This is especially
interesting for bug-reports and pull-requests/patchsets.

Today, I was looking into the issue, planning to implement it, and found
an oddity in our interface design. Naturally, the "create-message" tool
has to support multiple parents for the new commit. Currently, (as of
1e615437fafb3a1f60e9d079ca545b2329b6ae54) it takes the (one) parent-hash
and a tree-init-hash as arguments. Both are optional. As the
documentation explains:

> If <parent-hash> and <tree-init-hash> are not provided, we create a
> new issue.
>
> If <parent-hash> is provided and <tree-init-hash> is not, we create a
> new issue which is attached to an existing commit. That can be used
> (for example) to attach a bug report to a commit.
>
> If both are provided, we create a reply to an existing issue
> discussion.

It is obvious that the tool cannot simply be extended to accept multiple
parent-hashes, as the tree-init-hash is optional. Reverting the order
does also not work for the same reason.

Feeding additional parents via a '-p' option seems ugly, since there is
no sensible reason to differentiate between the first and any othre
parents in this situation. Hence, I suggest to have the tree-init-hash
be specified via an '-i' option.

This change would clearly affect the "new" and "reply" commands, as well
as "tag". Hence, they need adjustment if we decide to carry out the
changes I suggested in this message.

Signed-off-by: Julian Ganz <neither@nut.email>
Dit-status: open