Re: Allow replies to have additional references
> We might need to come up with a solution for these kind of rewrites of
> the history.
The solution is simple: re-issue the original reply, referencing the new
head. I did not do this with the latest reply, but I could add another
one.

When we are not interested in keeping that specific history, we can
easily drop the reference referencing the specific branch of the
discussion containing the reference.

> Also, we cannot do a out-of-repo DIT with additional references or we
> might want some way to pull discussions from the DIT repository into the
> code repository automatically if one refers from one to another.
I had the impression that we already had a discussion about this. We can
keep "pure" issues in another repository and import PRs. Alternatively,
we can simply accept code in the issue repository.