Catch empty subcommands
If we don't catch empty subcommands, we end up with
`handle_unknown_subcommand()` trying to invoke a program named
`git-dit-`. This, in turn, results in a "No such file or directory"
error, which is not especially useful for the user.

This commit solves the issue. Empty subcommands are handled by printing
the usage to stderr.