Only support rustc >= 1.15
We do not support rustc versions before 1.15, because we depend (not
directly, but in the chain) on "unicode-bidi" (as an example), which
itself does not support prior versions.

We think that supporting older versions has no benefit for us.
Distributions which ship rustc in an old version are welcome to use
tools like rustup or alternative package managers to pull in a new rustc
version to compile this package.

It is not our job to provide backports.