Fix Trailers iterator
Instead of iterating over all lines, we iterated only over one element
when scanning blocks. This lead to the algorithm only see a block of
text or trailers as a sequence of one-line blocks.

The resulting behaviour was the `Trailers` type not dumping lines which
looked like trailers but were embedded in blocks of texts. This commit
fixes the issue by properly iterating over all lines (until the next
blank one).