Redefine retrieval of issue references
The way issue references are currently retrieved relies on a single
function for each possible variant of reference type and location (e.g.
local or remote). The amount of functions necessary to fulfill the needs
of users is unacceptably large. Hence, we have to redesign the interface
for retrieval of references.