Metadata accumulation and targeted metadata retrieval
Within an issue, metadata is stored in trailers within messages.
Currently, a user is able to query a list of trailers present in a range
of messages, e.g. the issue's "head" branch. However, we do not yet
provide a facility for accumulating metadata.

This patchset introduces such facilities in the library and aims at
providing at least some of the functionality through the command line
tool.