Improve lifetimes of results of Issue related functions
For some of the functions for the `Issue` type, we missed annotating the
return types with an explicit lifetime. This resulted in the default
lifetime assigned through lifetime elision. Those were far too strict
since the result's lifetime was those of the `Issue` rather then that of
the underlying `git2::Repository`.

This patch set aims at improving the situation by adding explicit
lifetimes for various results of issue related functions.