Make TrailerValue::append() perform append in-place
A function named `append()` should always act in-place instead of
consume a value. Especially in this case, the design of the function
wasn'r a good idea in the first place: a `TrailerValue` will almost
always be encountered as part of a `Trailer`. Requireing callers to
re-assing the value whenever it is modified is just bad interface
design.