Open the repository in the individual subcommands
We currently open the repository in the main function. This has the
serious drawback that git-dit will look for a repository and abort on
failure even if the subcommand invoked does not require a repository.
Such subcommands include "check-message".

This patch-set resolved the problem by moving the responsibility of
opening a repository to the subcommands requiring it.