Remote priorization
Currently, all references available for an issue are treated equally in
some regards. For example, when querying the metadata, the next best
head-reference is used. This behavior is not optimal, especially in the
light of multiple remotes and upcoming functionality.

This patch-set aims at adding functionality for ordering references
regarding their associated remotes. The local reference will always have
preference. The order in which remote references are considered will be
settable by the user via a git-config variable.