Fix garbage collection of leaves considering remote refs
If remote references are considered during garbage collection, those
references should serve as starting points for the collection. We
erroneously added local references to the list of references (once
again) to be considered for collection instead. This resulted in leaves
not being collected when they should be.

This patch resolves the bug by pushing the remote references to the
correct revwalk.