Return set of unique issues from RepositoryExt functions
`RepositoryExt::issues()` and `RepositoryExt::issues_with_prefix()` both
return an iterator over issues. In both cases, the issues are not
guaranteed to be unique and with the current implementation, an issue
may well be reported twice.

We implicitly assumed that issues would be reported only once. This lead
to `git-dit-list` reporting an issue twice if both a remote and a local
head reference exist for that issue. `git-dit-gc` would try to remove a
reference twice and hence display an error.

We could fix the behavior in places where we made the assumption.
However, when querying a list of issues it is sensible to assume that
the list would contain unique issues. This patch-set therefore aims at
changing the aforementioned query functions of `RepositoryExt`, making
each of them return a set of issues.