Make git-dit-tag work without local head being present
The tag subcommand required a local head reference to be present for
operation. If no reference was present, the subcommand would produce an
error. This behavior was chosen because we did lack a selection
mechanism for remote heads. However, such a mechanism was recently
introduced.

This patch makes the "tag" subcommand use the remote priorization
mechanism for selecting a head reference.