Re: Issue Namespaces
> Things that we should implement then:
>
> 1. Moving things between namespaces - complicated from a data storage
> point of view, as moving alters hashes

If namespaces are realized solely via references, moving an issue is
unproblematic in this regard as the commits are not touched.

However, consistency of an issue's namespace across multiple
repositories is a problem. Pushes from users and replies may lead to
segmentation of issues or, in the worst case, duplication of entire
issues.

Such problems can be hidden from the user to some extend by
automatically "collapsing" or deduplicating issues (e.g. we only use the
id to identify issues and ignore the namespace for most purposes).
However, the storage itself will still be a mess.

For now, I'd opt for "passive" support for namespaces: issues are only
identified by their id but are still found if they appear within a
namespace. However, the namespace information is neither exported nor do
we offer tools for moving an issue from one namespace to another. In
fact, as of v0.3 git-dit may already provide complete "passive" support.