Redefine internals of Issue
Some existing and future functions of `Issue` operate on the initial
message of the issue. If we store only the `Issue`s id, we have to
retrieve the commit on multiple occasions, e.g. we have a potential for
a lot of overhead.

Ideally, we would store the initial message as a `git2::Commit`.
However, this is not practical due to some drawbacks of the interface.
Storing an `Object` on the other hand provides at least some caching as
well as other benefits.