Make author and committer settable for some subcommands
Currently, the author and committer of a new commit are taken from the
local configuration, e.g. the "git-config" of the local repository. For
the use-case of a developer creating issues from the command line, this
is perfectly acceptable.

For some other use-cases, however, the author and committer of new
messages should be settable explicitly via the command line. Affected
subcommands would gain new command-line parameters, e.g. `--author` and
`--committer`. The following subcommands may be extended:
 * new
 * reply
 * tag
 * create-message