Pre-collect names of local References in push_impl
With git2-0.6.7, the signature of the `References::names` changed. The
function now takes a mutable reference and imposes an additional
lifetime constraint on the returned iterator. The latter can not
out-life the `References` struct any more. This lead to compilation
failures.

The problem is resolved by pre-collecting all names as `String`s within
a closure.