Removal of trailing blank lines
In the library, we provide line oriented processing of text, e.g.
messages.

Various processing may yield iterators with blank lines at the beginning
or at the end of messages. While the former is relatively easy to filter
out using `Iterator::skip_while()`, the latter is not as easy to avoid
using rust standard tools.

In the library, we currently avoid the problem by removing trailing
whitespace from strings, which are collected from lines anyway. However,
library users may still be interested in a tool for removing trailing
blank lines.

This patch-set introduces an iterator adapter for performing this task.