Implement string-splitting for multi-word configurations
The problem fixed with this patch was the following:

If the `$PAGER` variable was not a one-word string, but for example
something like `less -R`, the previous source wanted to call
`Command::new("less -R")`, which does not work.

The patch changes the code to split the fetched configuration for the
command to be executed by whitespace, applying the first yielded word to
`Command::new()` and all following as args to the new `Command`
instance.