Remove one type parameter in trait with associated type
This patch removes one level of generic by transforming it to an
associated type.

This improves usability of the traits because of fewer generics.
Also, generic output values are problematic because every function which
wants to be generic over the trait has also to be generic over the
generic return type of the functions of the trait (see rust
documentation on associated types).

This patch improves this.