Yank git-dit from tree, leaving behind libgitdit and docs
Up until now, both the `git-dit` tool and the library `libgitdit`
hosting the actual implementation were developed in a single repository.
This approach was chosen in order to speed up the tool's development:
we would not have to wait for a release of the library but could use
features we just implemented directly. However, we planned to split
binary and library into separate trees early in the development already.

We intended to introduce a few more features before doing the split.
However, development of `git-dit` as a whole died down in the last
couple of years. We got a bit of motivation to work on this project
again and are planning something like a general overhaul. Since we, more
or less, go for a clean cut anyway, splitting the library now makes more
sense than postponing the split any longer.

This change only removes the `git-dit` tool. A future change will move
the library to the tree's top.