Move libgitdit to the tree's top
Up until now, both the `git-dit` tool and the library `libgitdit`
hosting the actual implementation were developed in a single repository.
However, we decided to split the executable and library. For this
purpose, we recently removed `git-dit` from the tree, leaving behind
conceptual documentation and the library.

This change now moves the library to the top of the tree. I.e., from now
on, we'll develop just the library in this tree.