message: use ? instead of try! in LineIteratorExt::check_message_format
We started developing this library waaay back, which shows in our error
handling in "a few" places. This change drops a `try!`, shutting up a
warning.

Needless to say, the error handling in this particular is still not
particularly ideomatic. We may come back rewriting the thing in the
future.