issue: fix Issue::messages
In the past, we could push all references, remote and local, associated
with an issue with a single glob starting with `**`. However, that
functionality appears to be broken. `Issue::all_refs` is probably also
broken but currently it's not covered by any tests and given that we
plan quite intrusive changes which will probably also affect that fn we
are going to ignore that for now.

Actually, I myself may have been the one breaking `Revwalk::push_glob`,
which means I'm obligated to track this regression down. But even if we
fix our dependency, I'd argue thet that glob was a bit fishy anyway (it
may well have selected references completely unrelated to git-dit) and
nothing we should rely upon.