error: add new Kind type for future custom Error type
Currently, this library relies on the `git2` crate for accessing git
repositories. In the future, however, we want to enable our users to
choose between that and other git implementations (such as the
`git-repository` crate). A recurring pattern will be traits with one
implementation for each library we support.

Many functions may fail in one way or another and failures may often
originate from the git library. Our current `Error` can therefore wrap
a `git2::Error`. If we want to support additional libraries, our `Error`
needs the ability to wrap error types specific to those libraries. At
the same time, we neither need nor want to support mixing git libraries.

We decided to make `Error` generic over the "inner" `Error` type.
Unfortunately, `error_chain` makes this hard (impossible?) to do. We
could simply use `Box<dyn Error>` for inner errors but that's not a
thing we want to do in library code. And it doesn't solve the problem of
library specific types for object IDs and references anyway.

So we'll need to roll our own, handwritten `Error`. And if we're gonna
do that anyway, we may as well name our `ErrorKind` simply `Kind` (since
it resides inside a module named `error`). Since we're using a new name,
we can introduce the new type without running into conflicts.

This change does just that. Future changes will switch out the existing
`Error` itself.